+++ packages/editor-ng/src/public-api.ts
@@ -1,3 +1,3 @@
 export * from './workspace.component';
 export * from './entry-renderer.directive';
-
+export * from './diff.component';